requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
django-cors-headers==3.11.0
pyjwt==2.6.0
requests==2.22.0
requests-unixsocket==0.2.0
django-environ == 0.9.0
gunicorn == 20.1.0  
reportlab == 3.6.10
cryptography == 37.0.4
boto3 == 1.26.73
web3 == 6.5.0
qrcode == 7.3.1